separate functions into 2 classes
