+++ .github/workflows/daily-static-gtfs-check.yaml
@@ -31,11 +31,17 @@ jobs:
         run: java -cp build/classes/java/test:build/libs/gtfu.jar gtfu.test.LoadAgencyDataTest -c src/main/resources/conf/cache -u ${{ secrets.LIVE_AGENCIES_URL }}
         env:
           SENDGRID_API_KEY: ${{secrets.SENDGRID_API_KEY}}
-      - name: Update Feed Data
+      - name: Update trip names
         working-directory: gtfu
-        run: java -cp build/libs/gtfu.jar gtfu.tools.UpdateFeedData -u ${{ secrets.LIVE_AGENCIES_URL }}
+        run: java -cp build/libs/gtfu.jar gtfu.tools.UpdateTripNames -u ${{ secrets.LIVE_AGENCIES_URL }}
         env:
           GH_ACCESS_TOKEN: ${{secrets.GH_ACCESS_TOKEN}}
+          SENDGRID_API_KEY: ${{secrets.SENDGRID_API_KEY}}
+      - name: Check feed expiration
+        working-directory: gtfu
+        run: java -cp build/libs/gtfu.jar gtfu.tools.CheckFeedExpiration -u ${{ secrets.LIVE_AGENCIES_URL }} -s -e
+        env:
+          GRAAS_BOT_WEBHOOK: ${{secrets.GRAAS_BOT_WEBHOOK}}
           SENDGRID_API_KEY: ${{secrets.SENDGRID_API_KEY}}
       - name: Generate Block Data
         working-directory: gtfu
